+++ src/lib_proj.cmake
@@ -117,7 +117,6 @@ SET(SRC_LIBPROJ_PROJECTIONS
         projections/wag7.cpp
         projections/lcca.cpp
         projections/geos.cpp
-        projections/etmerc.cpp
         projections/boggs.cpp
         projections/collg.cpp
         projections/comill.cpp
